requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
lancedb==0.4.3
pandas==2.2.1
openpyxl==3.1.2
pyarrow==15.0.0 